requirements.txt
# local package
-e .

# external requirements
click
Sphinx
coverage
python-dotenv>=0.5.1
typing
hydra-core
wandb
scikit-learn
torch
numpy
optuna
transformers
pillow
pandas
datasets
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.12.7
charset-normalizer
colored==1.4.4
Faker==17.6.0
frozenlist==1.3.3
idna==3.4
multidict==6.0.4
python-dateutil==2.8.2
requests==2.28.2
six==1.16.0
tenacity==8.2.2
termcolor==2.2.0
urllib3==1.26.14
yarl==1.8.2
nltk
matplotlib
chardet
pymediawiki
ruff
cchardet
umap-learn
networkx
wikidata
